feat: replace RPC provider from Alchemy to Infura DIN (#451)
* feat: replace alchemy by infura din

* chore: change secrets from VOYAGER to DIN for rpc calls

* chore: lint+fix

* chore: .env.example

* chore: test import fix

* chore: reduced coverage on old test to 50%

* chore: handle comments

* chore: fix comments

* chore: handle comment

* chore: change ALCHEMY to DIN in CI/CD

* chore: fix comments
