+++ .github/workflows/publish-npm.yml
@@ -73,7 +73,7 @@ jobs:
           VERSION: ${{ needs.prepare-deployment.outputs.VERSION }}
           LOG_LEVEL: ${{ needs.prepare-deployment.outputs.LOG_LEVEL }}
           STARKSCAN_API_KEY: ${{ secrets.STARKSCAN_API_KEY }}
-          ALCHEMY_API_KEY: ${{ secrets.ALCHEMY_API_KEY }}
+          DIN_API_KEY: ${{ secrets.DIN_API_KEY }}
       - name: Cache Build
         uses: actions/cache@v3
         id: cache
