+++ packages/starknet-snap/.env.example
@@ -9,9 +9,9 @@ VOYAGER_API_KEY=
 # Description: Environment variables for API key of STARKSCAN
 # Required: false
 STARKSCAN_API_KEY=
-# Description: Environment variables for API key of ALCHEMY
+# Description: Environment variables for API key of INFURA DIN
 # Required: false
-ALCHEMY_API_KEY=
+DIN_API_KEY=
 # Description: Environment variables for Log Level, 0 does not log anything, 6 logs everything
 # Possible Options: 0-6
 # Default: 0
