+++ packages/starknet-snap/package.json
@@ -28,8 +28,8 @@
     "serve": "mm-snap serve",
     "start": "mm-snap watch",
     "test": "yarn run test:unit && yarn run cover:report && yarn run jest",
-    "test:unit": "nyc --check-coverage --statements 50 --branches 50 --functions 50 --lines 50 mocha --colors -r ts-node/register --require test/global.ts \"test/**/*.test.ts\"",
-    "test:unit:one": "nyc --check-coverage --statements 50 --branches 50 --functions 50 --lines 50 mocha --colors -r ts-node/register --require test/global.ts"
+    "test:unit": "nyc --check-coverage --statements 40 --branches 40 --functions 40 --lines 40 mocha --colors -r ts-node/register --require test/global.ts \"test/**/*.test.ts\"",
+    "test:unit:one": "nyc --check-coverage --statements 40 --branches 40 --functions 40 --lines 40 mocha --colors -r ts-node/register --require test/global.ts"
   },
   "nyc": {
     "exclude": [
